tried everything...nothing is working
can revert back to old stuff if you want but professors code doesn’t
even work. I think the problem is we have a post method and ajax on the
same thing and it just will never work. keep trying, but it might not
be able to be done. I found so many errors in the initial stuff we had,
but no matter how much I fix, nothing ever works