debug print statements
I added print statements to try and debug, but they are all printing as
expected.  The register button is still not enabling when it should be
enabling. can one of you guys take a look? I must be missing something
glaringly obvious but I've totally got box-vision with this script now.